screencontrol: fix memory leak. [1/1]
PD#SWPL-146126

Problem:
1.the AMediaCodec didn't delete.
2.the buffer of recording canvas buffer's address
didn't free.

Solution:
1.call AMediaCodec_delete function when
stop record
2.when the buffer of recording canvas
buffer's address is given to encoder,
it will be freed

Verify:
T7c

+++ frameworks/services/screencontrol/include/ulit.h
@@ -34,7 +34,7 @@ private:
 inline int64_t getNowTimesUs() {
     struct timespec now;
     clock_gettime(CLOCK_BOOTTIME, &now);
-    int64_t now_time = now.tv_sec * 1000 * 1000 + now.tv_nsec / 1000;
+    int64_t now_time = (int64_t)now.tv_sec * 1000 * 1000 + (int64_t)now.tv_nsec / 1000;
     return now_time;
 }
